rag_factory/requirements.txt
youtube-transcript-api==0.6.2
python-dotenv==1.0.1
requests==2.32.3
openai==1.40.6
google-generativeai==0.7.2
supabase==2.6.0
